+++ test/tx-data/candidacy-declare.test.js
@@ -1,11 +1,10 @@
 import {DeclareCandidacyTxData} from '~/src';
-import {clearData} from '~/test/utils';
 
 describe('DeclareCandidacyTxData', () => {
     const txParamsData = {
         address: 'Mx7633980c000139dd3bd24a3f54e06474fa941e16',
         publicKey: 'Mpf9e036839a29f7fba2d5394bd489eda927ccb95acc99e506e688e4888082b3a3',
-        commission: 10,
+        commission: '10',
         coin: 'MNT',
         stake: '100',
     };